Add files via upload
One of the most important tasks in financial markets is to analyse historical returns on various investments. To perform this analysis, we need historical data for the assets. 

At this present time Bitcoin(BTC), Ethereum(ETH), Binance Coin(BNB), XRP(XRP) and Cardona(ADA) are the largest cryptocurrency by market cap.

The objective of this project is to calculate the potential risk and returns and to optimise a portfolio based upon the efficient frontier.